+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Support rekeying in `@algo-builder/runtime`
 
+### Bug Fixes
+
+- `Runtime` env is enforcing address use in ItxnField already `exist`. However, those addresses used in `acfg` or create asset transactions the addresses use for config may not exist. We fixed it in this [PR](https://github.com/scale-it/algo-builder/pull/550). Reported by @patrick
+
 ## v3.0.0 2021-12-22
 
 ### Improvements